Refactor UnsupportedPreparedStatementException
