+++ shardingsphere-proxy/shardingsphere-proxy-frontend/shardingsphere-proxy-frontend-core/src/main/java/org/apache/shardingsphere/proxy/frontend/exception/ExpectedExceptions.java
@@ -42,7 +42,6 @@ public final class ExpectedExceptions {
         EXCEPTIONS.add(ShardingSphereConfigurationException.class);
         EXCEPTIONS.add(SQLParsingException.class);
         EXCEPTIONS.add(DistSQLException.class);
-        EXCEPTIONS.add(UnsupportedPreparedStatementException.class);
     }
     
     /**
