+++ shardingsphere-proxy/shardingsphere-proxy-frontend/shardingsphere-proxy-frontend-core/src/test/java/org/apache/shardingsphere/proxy/frontend/exception/ExpectedExceptionsTest.java
@@ -38,7 +38,6 @@ public final class ExpectedExceptionsTest {
         assertTrue(ExpectedExceptions.isExpected(SQLParsingException.class));
         assertTrue(ExpectedExceptions.isExpected(DistSQLException.class));
         assertTrue(ExpectedExceptions.isExpected(NoDatabaseSelectedException.class));
-        assertTrue(ExpectedExceptions.isExpected(UnsupportedPreparedStatementException.class));
     }
     
     @Test
